Add Muen hypervisor detection code
The detection mechanism checks if the sinfo page magic is present. If
the Muen SK/hypervisor is detected, override IRQ and PCI init functions
and clear the TSC CPU feature since accessing the Timestamp counter is
not allowed on this platform.

To avoid unnecessary traps due to TSC access get_cycles() is aliased to
the muen_sched_clock_read() function.