Implement IRQ affinity handling
Let CPU cores store the device IRQs of associated devices in a global
list accessible to all CPU cores. This is required because the setup of
a device can happen on any core (missing the IRQ information in its
sinfo as the IRQ might not be assigned to this core).

Adjust muen-msi driver to use the new muen_smp_get_irq_affinity function
to retrieve the CPU core which is responsible to handle IRQ(s) of a
given device.