Also store event vectors in CPU affinity list
Let each core store unique target event vectors in the global affinity
list. This is required for the same reason as for device IRQs: An event
target might only be assigned to one subject of an SMP Linux, therefore
the core performing the setup of the associated virtual device does not
know about it.

Adjust the muen-input driver to use the muen_smp_get_event_vector
function to acquire the input vector from the CPU affinity list.