hvc_muen: Use muen_smp_one_match function
Also pin driver to the CPU of the associated channel event. This is
required because hvc_muen_put() runs with IRQs disabled, so it is not
possible to use smp_call_function_single() or smp_call_on_cpu()
functions to trigger an event on a remote CPU.

The initcall must be changed from console_initcall to early_initcall to
ensure that the SMP affinity db is ready.