muen: Override irq_enable in Muen IRQ chips
This allows on-demand setup of the per-cpu vector_irq descriptor.

Also introduce new smp_one_match_func to retrieve one entry from the
SMP affinity db by using a match function. Use this function in the
existing smp_one_match function.

Retrieve IRQ CPU affinity in muen_irq_free_desc and use it to set the
descriptor to VECTOR_UNUSED. The assumption that the request to do this
always originates from the CPU owning the IRQ is not true.