Remove printk from muensk_[mask|unmask]
While it is generally safe to call printk() in any context (according to
the Linux kernel docs), caution must be taken if the produced log
messages are directly written to the console because of the console log
level, and even worse if the console is a slow serial line. IMHO both
aspects are true in our case atm.

Therefore drop printk() calls in muensk_mask and muensk_unmask.