irq-muensk: Set the number of IRQs to get 1:1 mapping
The default value for nr_irqs/NR_IRQS is 64 in our kernel (it's not a
configuration option but defined via headers).  With CONFIG_SPARSE_IRQ,
the maximum number of supported IRQs by the kernel is actually higher
(NR_IRQS+8192).  Whenever a mapping is created, the kernel first tries
to allocate it in the current bitmap up to nr_irqs, so for values above,
there won't be a 1:1 mapping.  Only if no values are free, will the
kernel increase nr_irqs, but only does so as necessary (for single IRQs
just by one).

So to get a 1:1 mapping between virtual and physical IRQ also for
values > 64, we increase nr_irqs to the number of IRQs we support.
The kernel does query the architecture for the number of IRQs via
`arch_probe_nr_irqs()`, which is currently not implemented for arm64.
So that would be an option, however, we don't have access to the device
tree node there.  So doing this when the IRQ chip is initialized instead,
which happens shortly afterwards, will later allows us to read the value
from the device tree/policy.

Note that the size we configure when creating the IRQ domain does not
affect nr_irqs, it only defines the number of mappings the domain itself
can handle (with the linear domain this directly translates to an array
of the given size, with the tree based domain there wouldn't be an upper
limit).